Adds RRF_HOST_BUILD and fixes Hangprinter support
 - Only call serialUSB if SUPPORT_USB
 - Initializes some previously uninitialized variables.
 - Adds the missing Hangprinter registration so M669 K6 maps to a real kinematics.
   Instantiating the static descriptor and Create() helper now lets Move::SetKinematics construct HangprinterKinematics.
 - Fix warning in HangprinterKinematics.
   Replaces the variable-length stack buffers in CartesianToMotorSteps with HANGPRINTER_MAX_ANCHORS-sized arrays
   so GCC no longer flags zero-length writes when numAnchors is 0.
 - Implements new flex compensation for Hangprinter
 - Hangprinter: new quadratic forward transform
 - Hangprinter: all anchors drives control of all moves
   Hangprinter kinematics now advertises that all anchor drives control any Cartesian axis during normal moves, so async axis allocation includes motor C even when Z isn't mentioned.
   This mirrors the delta/CoreXY behavior.
 - Cache M569 values for expansion boards on main board.
   Hangprinter needs this for its can communication with ODrive on a separate can bus
 - Make M569.4 P<can_addrs> T<forces> allow a list of forces
   ... At least for Hangprinter
 - Hangprinter: Removes the M669 K-anchor to avoid name clash
 - Hangprinter: allow board address up to 49.0

+++ src/Movement/Move2.cpp
@@ -699,7 +699,7 @@ GCodeResult Move::ConfigureDriverBrakePort(GCodeBuffer& gb, const StringRef& rep
 
 	if (!seen)
 	{
-		reply.printf("Driver %u uses brake port ", driver);
+		reply.printf("Driver %zu uses brake port ", driver);
 		brakePorts[driver].AppendPinName(reply);
 # if SUPPORT_BRAKE_PWM
 		if (brakeVoltages[driver] < FullyOnBrakeVoltage)
@@ -1217,7 +1217,7 @@ void Move::ReportM569Parameters(size_t drive, const StringRef& reply) noexcept
 {
 	// Print the basic parameters common to all types of driver
 	reply.printf("Drive %u runs %s, active %s enable",
-					drive,
+					(unsigned int)drive,
 					(GetDirectionValue(drive)) ? "forwards" : "in reverse",
 					(GetEnableValue(drive) > 0) ? "high" : "low");
 #if SUPPORT_SLOW_DRIVERS
